fastapi_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic[email]
pydantic-settings
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
requests